Moved subrow into a seperate component
