+++ apps/app-orch/src/components/atoms/ApplicationProfileOverrideSubComponent/ApplicationProfileOverrideSubComponent.pom.ts
@@ -1,0 +1,18 @@
+/*
+ * SPDX-FileCopyrightText: (C) 2023 Intel Corporation
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { CyPom } from "@orch-ui/tests";
+
+const dataCySelectors = [] as const;
+type Selectors = (typeof dataCySelectors)[number];
+
+class ApplicationProfileOverrideSubComponentPom extends CyPom<Selectors> {
+  constructor(
+    public rootCy: string = "applicationProfileOverrideSubComponent",
+  ) {
+    super(rootCy, [...dataCySelectors]);
+  }
+}
+export default ApplicationProfileOverrideSubComponentPom;
