+++ apps/app-orch/src/components/organisms/setup-deployments/OverrideProfileValues/OverrideProfileTable.tsx
@@ -6,15 +6,14 @@
 import { adm, catalog } from "@orch-ui/apis";
 import { ApiError, Table, TableColumn, TableLoader } from "@orch-ui/components";
 import { InternalError, parseError, SharedStorage } from "@orch-ui/utils";
-import { MessageBanner } from "@spark-design/react";
-import { memo, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { useAppDispatch, useAppSelector } from "../../../../store/hooks";
 import {
   profileParameterOverridesSelector,
   setDeploymentApplications,
   setupDeploymentEmptyMandatoryParams,
 } from "../../../../store/reducers/setupDeployment";
-import ApplicationProfileParameterOverrideForm from "../../../atoms/ApplicationProfileParameterOverrideForm/ApplicationProfileParameterOverrideForm";
+import ApplicationProfileOverrideSubComponent from "../../../atoms/ApplicationProfileOverrideSubComponent/ApplicationProfileOverrideSubComponent";
 import "./OverrideProfileTable.scss";
 
 const dataCy = "overrideProfileTable";
@@ -28,40 +27,6 @@ export interface OverrideProfileTableProps {
   selectedPackage: catalog.DeploymentPackage;
   selectedProfile: catalog.DeploymentProfile;
 }
-
-interface ApplicationProfileOverrideSubComponentProps {
-  app: catalog.Application;
-  selectedProfile: catalog.DeploymentProfile;
-}
-
-/** This will render application profiles with override form upon row expand */
-const ApplicationProfileOverrideSubComponent = memo(
-  ({ app, selectedProfile }: ApplicationProfileOverrideSubComponentProps) => {
-    /** Application Profile configuration on the selected deployment package profile */
-    const appProfile = app.profiles?.find(
-      (profile) =>
-        // Check if the application profile name is what is seen selected in the selected deployment package profile
-        profile.name === selectedProfile?.applicationProfiles[app.name],
-    );
-
-    if (!appProfile) {
-      return (
-        <MessageBanner
-          messageTitle="Error while reading Parameter Template"
-          messageBody={`No profile found for app "${app.name}@${app.version} in Deployment Profile ${selectedProfile?.name}"`}
-          variant="error"
-        />
-      );
-    }
-
-    return (
-      <ApplicationProfileParameterOverrideForm
-        application={app}
-        applicationProfile={appProfile}
-      />
-    );
-  },
-);
 
 const OverrideProfileTable = ({
   selectedPackage,
